requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
Pillow==10.4.0
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.0.1
requests==2.31.0 